Added test to SLURM
