+++ bout_runners/config/logging_config.yaml
@@ -4,7 +4,7 @@ disable_existing_loggers: False
 formatters:
   simple:
     format:
-      "[%(asctime)s][%(levelname)-8s][%(filename)-22s][l%(lineno)-4s]
+      "[%(asctime)s][%(levelname)-8s][%(filename)-29s][l%(lineno)-4s]
       %(indent)s%(message)s"
 
 handlers:
