--- bout_runners/executor/__init__.py
@@ -1,1 +1,0 @@
-"""Package containing executor modules."""
