+++ bout_runners/runner/run_group.py
@@ -7,13 +7,11 @@ from pathlib import Path
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
 
 from bout_runners.database.database_connector import DatabaseConnector
-from bout_runners.executor.bout_paths import BoutPaths
-from bout_runners.runner.bout_run_setup import BoutRunSetup
+from bout_runners.parameters.bout_paths import BoutPaths
+from bout_runners.parameters.bout_run_setup import BoutRunSetup
 from bout_runners.runner.run_graph import RunGraph
-from bout_runners.submitter.abstract_submitters import (
-    AbstractClusterSubmitter,
-    AbstractSubmitter,
-)
+from bout_runners.submitter.abstract_cluster_submitter import AbstractClusterSubmitter
+from bout_runners.submitter.abstract_submitter import AbstractSubmitter
 from bout_runners.submitter.submitter_factory import get_submitter
