--- tests/local/unit/executor/__init__.py
@@ -1,1 +1,0 @@
-"""Package containing the unit test for the executor package."""
