+++ tests/slurm/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the tests for SLURM runs."""
