+++ tests/slurm/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the SLURM integrations tests."""
