+++ tests/slurm/integration/runners/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the SLURM integration test for the runners package."""
