+++ tests/slurm/unit/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the slurm unit tests for bout_runners."""
