+++ tests/slurm/unit/runners/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the SLURM unit test for the runners package."""
