+++ tests/slurm/unit/submitter/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the slurm unit test for the submitter package."""
